+++ packages/rhyolite_sync/test/state_puller_test.dart
@@ -41,6 +41,9 @@ class _FakeStateCaller implements IStateSyncContract {
   @override
   Future<StateWipeResponse> wipeVault(StateWipeRequest request, {RpcContext? context}) =>
       throw UnimplementedError();
+  @override
+  Future<StatePurgeResponse> purgeVault(StatePurgeRequest request, {RpcContext? context}) =>
+      throw UnimplementedError();
 }
 
 class _FakeHistoryCaller implements IHistoryContract {